use context for shared todo state
