+++ src/components/ButtonUI.vue
@@ -1,0 +1,10 @@
+<script setup lang="ts"></script>
+
+<template>
+  <button
+    type="button"
+    class="relative items-center px-4 py-2 text-sm font-medium text-gray-700 transition-colors border border-gray-300 rounded-md bg-gray-50 hover:bg-gray-100 focus:border-indigo-500 focus:outline-none focus:ring-1 focus:ring-indigo-500"
+  >
+    <span><slot /></span>
+  </button>
+</template>
