+++ src/views/HomeView.vue
@@ -1,7 +1,12 @@
-<script setup lang="ts"></script>
+<script setup lang="ts">
+import { computed } from "vue";
+import { useTodo } from "../feat/todo/useTodo";
+
+const { state } = useTodo();
+const todoLength = computed(() => state.value.todos.length);
+</script>
 
 <template>
-  <main>
-    <div>No todo</div>
-  </main>
+  <div v-if="todoLength > 0">編集するtodoをクリックして選んでください</div>
+  <div v-else>+をクリックしてtodoを作成してください</div>
 </template>
